Remove `closeCombatTracker` hook and move code to `renderCombatTracker`
